+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 
 - ⭐ Added try catch block to handle exception for JSON.parse function
+- ⭐ Changed default origin to ps.pndsn.com